Former-commit-id: 3c4cfb5bd241fa7e326a37c5971e1779c7c195f9

+++ hq-server/src/main/java/org/hyperic/hq/inventory/domain/IdentityAware.java
@@ -1,6 +1,6 @@
 package org.hyperic.hq.inventory.domain;
 
 public interface IdentityAware {
-	public Integer getId();
-	public void setId(Integer id);
+	public Long getId();
+	public void setId(Long id);
 }
